amd_hsmp: Update PwrEfficiencyModeSelection message
PwrEfficiencyModeSelection message is updated to include get option
in the latest PPR. Earlier only set option was supported.
Bit 31 of the input argument represents whether its set/get message.
If it is get message, current mode is returned in the output arguments.
If it is set message, successfully set value is returned in the output arg.
These messages are categorised as SET_GET messages.

For backward compatibility point of view
(existing user space tool and library who pass response_sz = 0),
the SET_GET messages are checked only for maximum response_sz,
not the exact response_sz. So even if the old library and tool
pass response_sz = 0 for set operation, driver will not complain.